requirements.txt
drawsvg[all]~=2.0
numpy ~= 1.25.2
lxml ~= 4.9.3
pillow ~= 10.0.0
vpype ~= 1.13

coverage ~= 6.2
pytest ~= 6.2.5
pdoc == 8.0.1
Jinja2 == 3.0.3
beautifulsoup4 ~=4.10.0
